+++ ui_engine/towa-app/src/views/EditorTab.vue
@@ -6,6 +6,7 @@ import { usePageLoader } from '@/composables/usePageLoader'
 import { useAutoSave } from '@/composables/useAutoSave'
 import PageSidePanel from '@/components/editor/PageSidePanel.vue'
 import TranslationPanel from '@/components/editor/TranslationPanel.vue'
+import AiToolbar from '@/components/editor/AiToolbar.vue'
 
 defineOptions({ name: 'EditorTab' })
 
@@ -89,6 +90,10 @@ onBeforeUnmount(() => window.removeEventListener('keydown', onKeydown))
 
 <template>
   <div>
+    <Teleport to="#towa-top-toolbar" defer>
+      <AiToolbar />
+    </Teleport>
+
     <Teleport to="#towa-left-panel" defer>
       <PageSidePanel
         :pages="pages"